+++ src/Nav.js
@@ -14,6 +14,8 @@ import AppSetting from './pages/TabbarPages/me/pages/AppSetting'
 import AboutUs from './pages/TabbarPages/me/pages/AboutUs'
 import MyFollow from './pages/TabbarPages/me/pages/MyFollow'
 import Reminder from './pages/TabbarPages/me/pages/Reminder'
+import MeScreen from './pages/TabbarPages/me/MeScreen'
+import Logining from './pages/TabbarPages/me/index'
 
 import ClubDetail from './pages/TabbarPages/news/pages/ClubDetail'
 import EventDetail from './pages/TabbarPages/news/pages/EventDetail'
@@ -47,9 +49,11 @@ class Nav extends Component {
                 <Stack.Screen name="EventDetail"     component={EventDetail} />
 
                 {/* 我的頁 */}
-                <Stack.Screen name="AboutUs"        component={AboutUs}/>
+                <Stack.Screen name="AboutUs"         component={AboutUs}/>
                 <Stack.Screen name="MyFollow"        component={MyFollow}/>
                 <Stack.Screen name="Reminder"        component={Reminder}/>
+                <Stack.Screen name="MeScreen"        component={MeScreen}/>
+                <Stack.Screen name="Logining"        component={Logining}/>
                 {/* 設置頁 */}
                 <Stack.Screen name="MeSetting"     component={MeSetting} />
                 <Stack.Screen name="AppSetting"     component={AppSetting} />
